+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## July 16, 2018
+* Added `getSessionId()` to fetch the current session ID.
+* Updated iOS to v4.2.1 [release notes](https://github.com/amplitude/Amplitude-iOS/releases/latest)
+* Updated Android to v2.18.1 [release notes](https://github.com/amplitude/Amplitude-Android/releases/latest)
+
 ## November 27, 2017
 * Updated Android to v2.16.0 [release notes](https://github.com/amplitude/Amplitude-Android/releases/latest)